Remove hardcoded triangle red clr from frag shader
Now a vertex vec4 attribute is used as an input for the vertex shader
and a vertex vec4 varying is used as an output for the vertex shader and
as an input for the fragment shader.

This way, each vertex color is handed down from the vertex shader to the
fragment shader and the color between vertices could be different.

Setting a different color for each vertex will interpolate the colors
along all fragments of the rasterized primitive.