Create gl-utils .h and .c files
Move LoadShader and Init functions into gl-utils to make them reusable.
Refactor DrawTriangle into the more reusable DrawShape function.

The purpose of all this is to prepare the ground for being able to draw
different shapes other than triangles, like a square and a circle, for
example, while keeping the source code reusable.