Fix "make always builds binaries" issue
Makefile was refactored in order to prevent the issue where make was
always building the target binaries no matter what.

BINS is now a list of the names of the binary files that are generated
by calling "make all" or "make <shape>" where shape could be triangle,
square, etc. By having the exact file name as part of the BINS list,
make is now able to decide correctly if the TARGET specified needs to
recompiled or not.