Add C++ rendering engine initial implementation
rendix will take place over gl-utils C function library as a complete
C++ rendering engine. It is meant to be used for educational purposes
only.

A new rendix-triangle test application is also included so application
and engine development could grow together and testing can be done in an
easy way.