Remove not used vertices in the pixel quad
To draw a 1x1 monochrome font bitmap solid pixel, either 1 quad or 2
triangles are needed to draw a square. Since the quad primitive is not
avaialble in OpenGL ES 2.0, drawing 2 different triangles is necessary.
The triangles strip is a good candidate to draw adjacent triangles that
share 2 vertices, so only 4 vertices are sent to the GPU using the
glDrawArrays call.